+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [next]
 
 - BREAKING fix(): reflect NUM_FRACTION_DIGITS to toObject path data [#8909](https://github.com/fabricjs/fabric.js/pull/8909)
+- chore(TS): export util types [#8915](https://github.com/fabricjs/fabric.js/pull/8915)
+- chore(TS): change enums with types [#8918](https://github.com/fabricjs/fabric.js/pull/8918)
 - chore(TS): export gradient types
 - chore(lint) export filter colors and brushes types [#8913](https://github.com/fabricjs/fabric.js/pull/8913)
 - chore(lint) Add a rule for import type [#8907](https://github.com/fabricjs/fabric.js/pull/8907)